DC offset changes
DC offset is now a added/subtracted as a constant, though it may change over time (eg. when in plasmasphere). There is some experimentation with having DC offset as a function of plasma density. Included data from sweep on 2024-08-20, where density is, as well as SC potential, which is compared to P4.